bumped coverage and split test packages
